Added a new version of the original data file.
It is now obvious that the reason the dates were all off
was because of the differences in Excel date formats
from Windows to Mac.

Basically, there are two different formats, and neither of them
are exactly compatible. They are four years and a day different -- which
was making the entrants really young in my analysis.

It shouldn't actually change anything else.

	new file:   data/source/Retention Data v2_14JAN2011.txt